editor/libeditor/tests/test_bug638596.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=638596
-->
<head>
  <title>Test for Bug 638596</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <script src="/tests/SimpleTest/EventUtils.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=638596">Mozilla Bug 638596</a>
<p id="display"></p>
<div id="content">
  <input type="password" style="font-size: 0">
</div>
<pre id="test">
<script type="application/javascript">

/** Test for Bug 638596 **/
SimpleTest.waitForExplicitFinish();
SimpleTest.waitForFocus(function() {
  var i = document.querySelector("input");
  i.focus();
  sendString("test");
  is(i.value, "test", "The correct value should be stored in the field");
  SimpleTest.finish();
});

</script>
</pre>
</body>
</html>
